+++ src/main/java/com/ediro/persistence/BasketRepository.java
@@ -15,6 +15,6 @@ import com.ediro.domain.Basket;
 public interface BasketRepository extends CrudRepository<Basket, String>,CustomBascketRepository
 {
 	public Basket findOneByBook_bookCodeAndMember_memberID(BigInteger book_code,String MemberID); 
-	public Basket findOneByBook_bookCodeAndMember_mid(BigInteger book_code,int mid);
+	public Basket findOneByBook_bookCodeAndMember_mid(long book_code,int mid);
 	
 }
